requirements.txt
Django==5.2
djangorestframework==3.15.0
djangorestframework-simplejwt==5.2.2
django-filter==24.1
celery==5.5.0
redis==4.7.0
